Moved SearchQueryField into Search.
